Product testcode

+++ townMarket/src/main/java/com/example/townmarket/product/service/ProductServiceImpl.java
@@ -1,15 +1,17 @@
 package com.example.townmarket.product.service;
 
 import com.example.townmarket.commons.dto.PageDto;
-import com.example.townmarket.product.dto.ProductDto;
+import com.example.townmarket.product.dto.PagingProductResponse;
+import com.example.townmarket.product.dto.ProductRequestDto;
+import com.example.townmarket.product.dto.ProductResponseDto;
 import com.example.townmarket.product.entity.Product;
 import com.example.townmarket.product.repository.ProductRepository;
 import com.example.townmarket.user.entity.User;
+import java.util.List;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,13 +23,14 @@ public class ProductServiceImpl implements ProductService {
 
   @Override
   @Transactional
-  public String addProduct(User user, ProductDto productDto) {
+  public String addProduct(User user, ProductRequestDto productRequestDto) {
     Product product = Product.builder()
-        .productName(productDto.getProductName())
-        .productPrice(productDto.getProductPrice())
-        .productStatus(productDto.getProductStatus())
-        .productCategory(productDto.getProductCategory())
-        .productEnum(productDto.getProductEnum())
+        .productName(productRequestDto.getProductName())
+        .productPrice(productRequestDto.getProductPrice())
+        .productStatus(productRequestDto.getProductStatus())
+        .productCategory(productRequestDto.getProductCategory())
+        .productEnum(productRequestDto.getProductEnum())
+        .userId(user.getId())
         .build();
     productRepository.save(product);
     return "상품을 성공적으로 등록했습니다";
@@ -35,12 +38,13 @@ public class ProductServiceImpl implements ProductService {
 
   @Override
   @Transactional
-  public ProductDto showProduct(long productId) {
+  public ProductResponseDto showProduct(long productId) {
     Product product = productRepository.findById(productId).orElseThrow(
         () -> new IllegalArgumentException("존재하지 않는 상품입니다")
     );
 
-    return new ProductDto(
+    return new ProductResponseDto(
+        product.getId(),
         product.getProductName(),
         product.getProductPrice(),
         product.getProductStatus(),
@@ -51,40 +55,44 @@ public class ProductServiceImpl implements ProductService {
 
   @Override
   @Transactional
-  public Page<ProductDto> viewAllProducts(PageDto pageDto) {
-    return productRepository.findAll(makePage(pageDto))
-        .map(product -> new ProductDto(
-            product.getProductName(),
-            product.getProductPrice(),
-            product.getProductStatus(),
-            product.getProductCategory(),
-            product.getProductEnum()));
+  public Page<PagingProductResponse> viewAllProduct(PageDto pageDto) {
+    List<Product> products = findAllProduct();
+    return new PageImpl<>(
+        products.stream().map(PagingProductResponse::new).collect(Collectors.toList()),
+        pageDto.toPageable(),
+        pageDto.getSize());
   }
 
   @Override
   @Transactional
-  public String updateProduct(long productId, ProductDto productDto) {
+  public String updateProduct(long productId, ProductRequestDto productDto, User user) {
     Product product = productRepository.findById(productId).orElseThrow(
         () -> new IllegalArgumentException("존재하지 않는 상품입니다")
     );
-    product.update(productDto);
-    return "상품이 성공적으로 업데이트되었습니다";
+    if (user.getId().equals(productId)) {
+      product.update(productDto);
+      return "상품이 성공적으로 업데이트되었습니다";
+    } else {
+      throw new IllegalArgumentException("본인의 상품이 아닙니다");
+    }
   }
 
   @Override
   @Transactional
-  public void deleteProduct(long productId) {
+  public void deleteProduct(long productId, User user) {
     Product product = productRepository.findById(productId).orElseThrow(
         () -> new IllegalArgumentException("존재하지 않는 상품입니다")
     );
 
-    productRepository.deleteById(product.getId());
+    if (user.getId().equals(productId)) {
+      productRepository.deleteById(product.getId());
+    } else {
+      throw new IllegalArgumentException("본인의 상품이 아닙니다");
+    }
   }
 
-  // Pageable 생성 메서드
-  public Pageable makePage(PageDto pageDto) {
-    Sort.Direction direction = pageDto.isAsc() ? Sort.Direction.ASC : Sort.Direction.DESC;
-    Sort sort = Sort.by(direction, pageDto.getSortBy());
-    return PageRequest.of(pageDto.getPage() - 1, pageDto.getSize(), sort);
+  @Override
+  public List<Product> findAllProduct() {
+    return productRepository.findAll();
   }
 }
